feat(no-task-available): rename AllTasksCompleted folder to NoTaskAvailable
